+++ cdo_ansible/roles/cdo_encrypt_credentials/files/encrypt.py
@@ -5,16 +5,10 @@ import sys
 import time
 import json
 import base64
+from Crypto.PublicKey import RSA
+from Crypto.Cipher import PKCS1_v1_5
 
-def load_package(lib_test, package):
-    try:
-        __import__(lib_test)
-    except ImportError:
-        os.system(f"{sys.executable} -m pip install {package} --user")
-    
 def main(username, password, public_key):
-    from Crypto.PublicKey import RSA
-    from Crypto.Cipher import PKCS1_v1_5
     key = RSA.importKey(base64.b64decode(public_key))
     encryptor = PKCS1_v1_5.new(key)
     print(
@@ -31,6 +25,4 @@ if __name__ == "__main__":
     username = sys.argv[1]
     password = sys.argv[2]
     public_key = sys.argv[3]
-    load_package("Crypto.PublicKey", "pycryptodome")
-    time.sleep(5)
     main(username, password, public_key)
